requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
black==23.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cytoolz==0.12.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.23.5
packaging==23.0
parsimonious==0.9.0
pathspec==0.11.1
platformdirs==3.2.0
protobuf==4.22.1
py-solc-x==1.1.1
pycodestyle==2.10.0
pycryptodome==3.17
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.4
pywin32==306
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
rlp==3.0.0
semantic-version==2.10.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
toolz==0.12.0
tzdata==2022.4
urllib3==1.26.12
web3==6.0.0
websockets==10.4
yarl==1.8.2
